+++ packages/@cosmos-building-blocks/common/src/cdk-tool-kit-app.ts
@@ -1,5 +1,0 @@
-import { App } from '@aws-cdk/core';
-import { CDKToolKit } from './cdk-tool-kit';
-
-const app = new App();
-new CDKToolKit(app, 'CDKToolkit');